mysite/pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Александр Полицарнв <norilskijsam@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.0"
django-rest-framework = "^0.1.0"
django-filter = "^23.3"
django-debug-toolbar = "^4.2.0"
pillow = "^10.1.0"
gunicorn = "^21.2.0"
sentry-sdk = "^1.34.0"
drf-spectacular = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
